erro no props do CategoryList
